core(utils): Moved package according to source packaging
