--- src/test/java/com/shortthirdman/primekit/essentials/ArrayUtilsTest.java
@@ -1,4 +1,0 @@
-package com.shortthirdman.primekit.essentials;
-
-public class ArrayUtilsTest {
-}
