+++ src/test/java/com/shortthirdman/primekit/essentials/common/util/ArrayUtilsTest.java
@@ -1,0 +1,4 @@
+package com.shortthirdman.primekit.essentials.common.util;
+
+class ArrayUtilsTest {
+}
